* maintenance/3.2.x: (2223 commits)
  [UIMA-6377] Spurious multipleReferencesAllowed warning when serializing empty arrays
  UIMA-6367: JCas cover annotation created in PEAR context replaced during index operations
  UIMA-6367: JCas cover annotation created in PEAR context replaced during index operations
  Corrected a spelling error
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release uimaj-3.2.0
  [UIMA-6310] JCasClassLoaderTest sometimes fails
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release uimaj-3.2.0
  [UIMA-6305] Convert UIMAv3 User's Guide from DocBook to Asciidoc
  [UIMA-6305] Convert UIMAv3 User's Guide from DocBook to Asciidoc
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release uimaj-3.2.0
  [UIMA-6329] UIMA Java SDK 3.2.0 release
  [UIMA-6353] Logger_common_impl.error(String arg0, Object arg1, Object arg2) swallows last argument
  [UIMA-6329] UIMA Java SDK 3.2.0 release
  [UIMA-6352] reinit(CASCompleteSerializer) does not properly reset indexes
  [UIMA-6352] reinit(CASCompleteSerializer) does not properly reset indexes
  [UIMA-6329] UIMA Java SDK 3.2.0 release
  [maven-release-plugin] prepare for next development iteration
  ...

# Conflicts:
#	PearPackagingMavenPlugin/pom.xml
#	README
#	RELEASE_NOTES.html
#	aggregate-uimaj-docbooks/pom.xml
#	aggregate-uimaj-eclipse-plugins/pom.xml
#	aggregate-uimaj/pom.xml
#	jVinci/pom.xml
#	jcasgen-maven-plugin/pom.xml
#	pom.xml
#	uima-docbook-overview-and-setup/pom.xml
#	uima-docbook-references/pom.xml
#	uima-docbook-tools/pom.xml
#	uima-docbook-tutorials-and-users-guides/pom.xml
#	uimaj-adapter-soap/pom.xml
#	uimaj-adapter-vinci/pom.xml
#	uimaj-bootstrap/pom.xml
#	uimaj-component-test-util/pom.xml
#	uimaj-core/pom.xml
#	uimaj-cpe/pom.xml
#	uimaj-document-annotation/pom.xml
#	uimaj-eclipse-feature-runtime/pom.xml
#	uimaj-eclipse-feature-tools/pom.xml
#	uimaj-eclipse-update-site/pom.xml
#	uimaj-ep-cas-editor-ide/pom.xml
#	uimaj-ep-cas-editor/pom.xml
#	uimaj-ep-configurator/pom.xml
#	uimaj-ep-debug/pom.xml
#	uimaj-ep-jcasgen/pom.xml
#	uimaj-ep-launcher/pom.xml
#	uimaj-ep-pear-packager/pom.xml
#	uimaj-ep-runtime/pom.xml
#	uimaj-examples/pom.xml
#	uimaj-json/pom.xml
#	uimaj-parent/pom.xml
#	uimaj-test-util/pom.xml
#	uimaj-tools/pom.xml

+++ uimaj-adapter-vinci/src/main/java/org/apache/uima/adapter/vinci/VinciAnalysisEngineServiceStub.java
@@ -49,9 +49,15 @@ import org.apache.vinci.transport.VinciFrame;
 import org.apache.vinci.transport.context.VinciContext;
 import org.apache.vinci.transport.document.AFrame;
 
+/**
+ * The Class VinciAnalysisEngineServiceStub.
+ */
 public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub {
+  
+  /** The m vinci client. */
   private VinciClient mVinciClient;
 
+  /** The m owner. */
   private Resource mOwner;
   
   /**
@@ -64,6 +70,7 @@ public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub
    */
   private int mGetMetaDataTimeout;
 
+  /** The Constant debug. */
   private static final boolean debug = System.getProperty("DEBUG") != null;
   
   /**
@@ -73,11 +80,27 @@ public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub
   private static final List SUPPORT_XCAS_V1 = Collections.unmodifiableList(
           Arrays.asList(new String[]{"1"}));
   
+  /**
+   * Instantiates a new vinci analysis engine service stub.
+   *
+   * @param endpointURI the endpoint URI
+   * @param owner the owner
+   * @throws ResourceInitializationException the resource initialization exception
+   */
   public VinciAnalysisEngineServiceStub(String endpointURI, Resource owner)
           throws ResourceInitializationException {
     this(endpointURI, null, owner, null);
   }
 
+  /**
+   * Instantiates a new vinci analysis engine service stub.
+   *
+   * @param endpointURI the endpoint URI
+   * @param timeout the timeout
+   * @param owner the owner
+   * @param parameters the parameters
+   * @throws ResourceInitializationException the resource initialization exception
+   */
   public VinciAnalysisEngineServiceStub(String endpointURI, Integer timeout, Resource owner,
           Parameter[] parameters) throws ResourceInitializationException {
     mOwner = owner;
@@ -124,7 +147,7 @@ public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub
       
       //store timeout for use in later RPC calls
       if (timeout != null) {
-        mTimeout = timeout.intValue();
+        mTimeout = timeout;
       } else {
        mTimeout = mVinciClient.getSocketTimeout(); //default
       }
@@ -144,11 +167,14 @@ public class VinciAnalysisEngineServiceStub implements AnalysisEngineServiceStub
 
 
   /**
+   * Call get meta data.
+   *
+   * @return the resource meta data
+   * @throws ResourceServiceException the resource service exception
    * @see ResourceServiceStub#callGetMetaData()
-   * @return -
    */
   @Override
-public ResourceMetaData callGetMetaData() throws ResourceServiceException {
+  public ResourceMetaData callGetMetaData() throws ResourceServiceException {
     try {
       // create Vinci Frame
       VinciFrame queryFrame = new VinciFrame();
@@ -196,19 +222,23 @@ public ResourceMetaData callGetMetaData() throws ResourceServiceException {
   }
 
   /**
-   * @see AnalysisEngineServiceStub#callProcess(CAS)
-   * @param aCAS theCAS
+   * Call process.
+   *
+   * @param aCAS the a CAS
    * @throws ResourceServiceException the resource service exception
+   * @see AnalysisEngineServiceStub#callProcess(CAS)
    */
   @Override
-public void callProcess(CAS aCAS) throws ResourceServiceException {
+  public void callProcess(CAS aCAS) throws ResourceServiceException {
     doProcess(aCAS);
   }
 
   /**
-   * @param aCAS -
+   * Call process cas.
+   *
+   * @param aCAS the a CAS
+   * @throws ResourceServiceException the resource service exception
    * @see CasObjectProcessor#processCas(CAS)
-   * @throws ResourceServiceException -
    */
   public void callProcessCas(CAS aCAS) throws ResourceServiceException {
     doProcess(aCAS);
@@ -216,6 +246,9 @@ public void callProcess(CAS aCAS) throws ResourceServiceException {
 
   /**
    * The actual process call.
+   *
+   * @param aCAS the a CAS
+   * @throws ResourceServiceException the resource service exception
    */
   private void doProcess(CAS aCAS) throws ResourceServiceException {
     try {
@@ -247,18 +280,23 @@ public void callProcess(CAS aCAS) throws ResourceServiceException {
   }
 
   /**
+   * Destroy.
+   *
    * @see ResourceServiceStub#destroy()
    */
   @Override
-public void destroy() {
+  public void destroy() {
     mVinciClient.close();
   }
 
   /**
+   * Call batch process complete.
+   *
+   * @throws ResourceServiceException the resource service exception
    * @see CasObjectProcessor#batchProcessComplete(org.apache.uima.util.ProcessTrace)
    */
   @Override
-public void callBatchProcessComplete() throws ResourceServiceException {
+  public void callBatchProcessComplete() throws ResourceServiceException {
     try {
       // create Vinci Frame ( Data Cargo)
       AFrame queryFrame = new AFrame();
@@ -272,10 +310,13 @@ public void callBatchProcessComplete() throws ResourceServiceException {
   }
 
   /**
+   * Call collection process complete.
+   *
+   * @throws ResourceServiceException the resource service exception
    * @see CasObjectProcessor#collectionProcessComplete(org.apache.uima.util.ProcessTrace)
    */
   @Override
-public void callCollectionProcessComplete() throws ResourceServiceException {
+  public void callCollectionProcessComplete() throws ResourceServiceException {
     try {
       // create Vinci Frame ( Data Cargo)
       AFrame queryFrame = new AFrame();
@@ -290,9 +331,11 @@ public void callCollectionProcessComplete() throws ResourceServiceException {
   }
 
   /**
+   * Call is read only.
+   *
+   * @return true, if successful
+   * @throws ResourceServiceException the resource service exception
    * @see CasObjectProcessor#isReadOnly()
-   * @return -
-   * @throws ResourceServiceException -
    */
   public boolean callIsReadOnly() throws ResourceServiceException {
     try {
@@ -311,9 +354,11 @@ public void callCollectionProcessComplete() throws ResourceServiceException {
   }
 
   /**
+   * Call is stateless.
+   *
+   * @return true, if successful
+   * @throws ResourceServiceException the resource service exception
    * @see CasObjectProcessor#isStateless()
-   * @return -
-   * @throws ResourceServiceException -
    */
   public boolean callIsStateless() throws ResourceServiceException {
     try {
@@ -331,6 +376,12 @@ public void callCollectionProcessComplete() throws ResourceServiceException {
     }
   }
 
+  /**
+   * Call get supported X cas versions.
+   *
+   * @return the list
+   * @throws ResourceServiceException the resource service exception
+   */
   public List callGetSupportedXCasVersions() throws ResourceServiceException {
     try {
       // create Vinci Frame ( Data Cargo)
@@ -352,6 +403,7 @@ public void callCollectionProcessComplete() throws ResourceServiceException {
       throw new ResourceServiceException(e);
     }
   }
+
   /**
    * Gets whether socket keepAlive is enabled, by consulting the
    * PerformanceTuningSettings.  (If no setting specified, defaults
